refactor: accessToken 관리 함수 유틸로 분리
